Use single-precision float-point values instead of double-precision values whenever data is being handled that will be written into the property value of a config element.  The C++ code is expecting single-precision values, and rounding errors tend to be exacerbated when using double-precision values.  Some double-precision values are still used for internal calcuations.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15387 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vrjuggler/vrjconfig/customeditors/display_window/org/vrjuggler/vrjconfig/customeditors/display_window/SimulatorViewportCreateDialog.java
@@ -85,9 +85,9 @@ public class SimulatorViewportCreateDialog
       this.pack();
    }
 
-   public Double getVertialFOV()
+   public Float getVertialFOV()
    {
-      return (Double) mVertFOVField.getValue();
+      return (Float) mVertFOVField.getValue();
    }
 
    public Object getCameraPosition()
@@ -104,7 +104,7 @@ public class SimulatorViewportCreateDialog
    {
       boolean valid_fov, camera_set, wand_set;
 
-      valid_fov  = ((Double) mVertFOVField.getValue()).doubleValue() > 0.0;
+      valid_fov  = ((Float) mVertFOVField.getValue()).doubleValue() > 0.0;
       camera_set = (((ConfigElementPointer) getCameraPosition()).getTarget() != null);
       wand_set   = (((ConfigElementPointer) getWandPosition()).getTarget() != null);
 
@@ -149,7 +149,7 @@ public class SimulatorViewportCreateDialog
       mVertFOVLabel.setText("Vertical Field of View");
       mVertFOVField.setMinimumSize(new Dimension(50, 19));
       mVertFOVField.setPreferredSize(new Dimension(50, 19));
-      mVertFOVField.setValue(new Double(80.0));
+      mVertFOVField.setValue(new Float(80.0));
       mVertFOVField.setHorizontalAlignment(SwingConstants.TRAILING);
       mVertFOVField.setFocusLostBehavior(JFormattedTextField.COMMIT_OR_REVERT);
       mCameraPosLabel.setLabelFor(mCameraPosEditor);
